requirement.txt
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.4.0
bcrypt==4.2.0
black==24.4.2
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
genson==1.3.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
inflect==5.6.2
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pip==22.0.2
platformdirs==4.2.2
pluggy==1.5.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.3.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
rich==13.7.1
setuptools==59.6.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlalchemy-json==0.7.0
starlette==0.37.2
toml==0.10.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
